% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.r
\name{uniquity}
\alias{uniquity}
\title{Estimating biotic uniqueness of biological samples}
\usage{
uniquity(sstable, classes = NULL, weights = NULL, presabs = TRUE,
  rep = 100, size = 0.66, refspecies = 0, rem_single = FALSE)
}
\arguments{
\item{sstable}{a data.frame with with a species/site matrix (or an OTU table)
that has sites/samples as columns and species/OTUs as rows, and observations as
presence/absence data or abundance data.}

\item{classes}{a list (dataframe) assigning samples to a set of unique classes (e.g. habitat-types). Each class
represents an exclusive proportion af the investigated total area). (data.frame should contain these column names: site, class).
If no classes are supplied, each samples is assigned to a unique class, and these will receive equal weights.}

\item{weights}{a list (dataframe) with the weight (proportion of total) of each
class. (data.frame should contain these column names: class, weight).
If no weights are supplied, classes are assigned equal weights.}

\item{presabs}{TRUE/FALSE. reduce species/site matrix to presence absence (0/1). Default is TRUE}

\item{rep}{any positive integer. The number of replicates pr site evaluation.
(i.e. the number of times to construct a pool of reference sites for evaluating each site/sample). Default is 100.}

\item{size}{any positive number. The number of reference sites to construct
for counting the number of unique species/OTUs of the sample/site in focus (i.e the uniquity of the site/sample).
each each compare. If \code{size} < 1, the algorithm uses that proportion
  of the sites investigated as the number of reference sites. If \code{size} =>1, that absolute number
  of reference sites wll be constructed. Default is 0.66 (=66\% of samples/sites)}

\item{refspecies}{any positive integer or 0. The number of species/OTUs to draw to construct a 
reference site. If \code{refspecies} = 0,  the function to draw a random figure based on
the species richness and weight of investigated sites. Any other figure will 
result in a fixed number being drawn for each reference site.}

\item{rem_single}{TRUE/FALSE. remove species that only occurs in one site
before simulation. Default = FALSE}
}
\value{
Function \code{uniquity} returns a list of results based on the input
table and parameters.
  \enumerate{
  \item \code{uniquity_score} the simulated uniquity contribution of each site, relative to species richness.
  \item \code{adjusted_uniquity} - number of curated (parent) OTUs
  \item \code{uniquity_table} the final table with number of unique observations pr species per site
  \item \code{uniquity_species_scores} sum of contribution per species.
  \item \code{Summed_species_weights} sum of weight pr species (used for drawing species for reference sites).
  \item \code{Average_species_weight_pr_site} average species weight pr site (with a weight).
  \item \code{Top20_unique_sites} top 20 unique sites.
  \item \code{Top20_unique_species} top 20 species with highest uniquity impact.
  \item \code{removed_singleton_species} which species were removed (if rem-single is et to TRUE).
  \item \code{Accumulated_uniqueness} uniquity uniquity metric per replicate.
  \item \code{site_richness} species richness per site.}
}
\description{
This algorithm \code{uniquity} estimates the uniqueness of a set of communities
}
\examples{
uniquity(my_table)
uniquity(my_table, my_classes, my_weights, presabs = FALSE, rep = 50, size = 100, refspecies = 50, rem_single = TRUE)
}
\author{
Tobias Guldberg Frøslev
}
