% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{uniquity2}
\alias{uniquity2}
\title{Estimating biotic uniqueness of biological samples}
\usage{
uniquity2(sstable, classes = NULL, weights = NULL, presabs = TRUE,
  background_sites = 50, rep = 50, non_correspondence = TRUE,
  total_coverage = 0)
}
\arguments{
\item{sstable}{a data.frame with with a species/site matrix (or an OTU table)
that has sites/samples as columns and species/OTUs as rows, and observations as
presence/absence data or abundance data.}

\item{classes}{a list (dataframe) assigning samples to a set of unique classes (e.g. habitat-types). Each class
represents an exclusive proportion af the investigated total area). (data.frame should contain these column names: site, class).
If no classes are supplied, each samples is assigned to a unique class, and these will receive equal weights.}

\item{weights}{a list (dataframe) with the weight (proportion of total) of each
class. (data.frame should contain these column names: class, weight).
If no weights are supplied, classes are assigned equal weights.}

\item{presabs}{TRUE/FALSE. reduce species/site matrix to presence absence (0/1). Default is TRUE}

\item{background_sites}{any positive number. The number of background sites to construct
for counting the number of unique species/OTUs of the sample/site in focus (i.e the uniquity of the site/sample).
each each compare. If \code{background_sites} < 1, the algorithm uses that proportion
  of the sites investigated as the number of background sites. If \code{background_sites} =>1, that absolute number
  of background sites wll be constructed. Default is 0.66 (=66\% of samples/sites)}

\item{rep}{any positive integer. The number of replicates pr site evaluation.
(i.e. the number of times to construct a pool of background sites for evaluating each site/sample). Default is 100.}

\item{non_correspondence}{TRUE/FALSE. If set to TRUE non-matching entries (classes, weights, sites) will be removed from the dataet.
If set to FALSE, non non-matching entries will throw an error. Can be usefull for subsetting data, 
as subsetting will only be necessary in one part of the data (species-site table)}

\item{total_coverage}{any number from 0 to 1. The fraction of total area investigated. If set to 0 the supplied
weights will be used, unless the total weight exceeds 1, in wich case weights are scaled for a total of 1. If set to any
number above 0 (including 1), weights are scaled for a total weight of \code{total_coverage}.}
}
\value{
Function \code{uniquity} returns a list of results based on the input
table and parameters.
  \enumerate{
  \item \code{simple_site_score} the sum of inverse species weights, as a fast approximation of the estimated uniquity.
  \item \code{uniquity_score} the simulated uniquity contribution of each site, relative to species richness.
  \item \code{adjusted_uniquity} - number of curated (parent) OTUs
  \item \code{uniquity_table} the final table with number of unique observations pr species per site
  \item \code{uniquity_species_scores} sum of contribution per species.
  \item \code{Unique_species_avg} average of contribution per species for sites where it occurs.
  \item \code{Summed_species_weights} sum of weight pr species (used for drawing species for background sites).
  \item \code{Average_species_weight_pr_site} average species weight pr site (with a weight).
  \item \code{Top20_unique_sites} top 20 unique sites.
  \item \code{Top20_unique_species} top 20 species with highest uniquity impact.
  \item \code{Accumulated_uniqueness} uniquity uniquity metric per replicate.
  \item \code{site_richness} species richness per site.
  \item \code{replicates} The number of replicates used for each site uniquity estimation.
  \item \code{background_sites} The number of background sites sampled.
  \item \code{Removed_classes} classes removed due to non correspondence.
  \item \code{Removed_sites} sites removed due to non correspondence.
  \item \code{total_coverage} sum of site weights used in weighting of species.}
}
\description{
This algorithm \code{uniquity} estimates the uniqueness of a set of communities
}
\examples{
uniquity(my_table)
uniquity(my_table, my_classes, my_weights, presabs = FALSE, rep = 50, size = 100, non_correspondence = TRUE, total_coverage = 0.5)
}
\author{
Tobias Guldberg Frøslev
}
